Add support for eth_call() in Seth RPC (STL-1568)
The model here is straightforward:
 - Add a read-only message call
 - Fix the machinery for passing block_id to Burrow
 - Fix the machinery for TAG|QUANTITY elements
 - Send the transaction the normal way then poll it out

Includes a handful of drive-by fixes to minor errors and build problems
hand-ported from upstream.  These should be benign (and were required for the
copmonent to build against changing dependencies) but I make this note to avoid
surprising reviewers, especially as it's all in one big commit.

+++ processor/src/seth_tp/handler/sawtooth_app_state.go
@@ -209,8 +209,11 @@ func (s *SawtoothAppState) SetStorage(address crypto.Address, key, value binary.
 	return nil
 }
 
+// FIXME: Older versions of Burrow require int64 type in the function signature, as legacy into Tendermint,
+// but block numbers should really be unsigned throughoiut.  This needs to be corrected (and cast removed)
+// when we update the dependencies to a later core Burrow (STL-1657, STL-1658)
 func (s *SawtoothAppState) GetBlockHash(blockNumber int64) (binary.Word256, error) {
-	blockInfo, err := getBlockInfo(s.mgr.state, blockNumber)
+	blockInfo, err := getBlockInfo(s.mgr.state, uint64(blockNumber))
 	if err != nil {
 		return binary.Zero256, fmt.Errorf("Failed to get block info: %v", err.Error())
 	}
